# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MangaGramm API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /series/{seriesId}/chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters of a series
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create a new chapter
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChapterBody"
      responses:
        "201":
          description: Chapter created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/{chapterId}:
    get:
      operationId: getChapter
      tags: [chapters]
      summary: Get chapter detail with pages
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapter detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterDetail"
    put:
      operationId: updateChapter
      tags: [chapters]
      summary: Update a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChapterBody"
      responses:
        "200":
          description: Chapter updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapter deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /chapters/{chapterId}/pages:
    post:
      operationId: addPages
      tags: [pages]
      summary: Add pages to a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPagesBody"
      responses:
        "201":
          description: Pages added
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Page"
...
  /discover/trending:
    get:
      operationId: getTrending
      tags: [discover]
      summary: Get trending series
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Trending series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Series"
  /discover/latest-updates:
    get:
      operationId: getLatestUpdates
      tags: [discover]
      summary: Get latest chapter updates
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Latest updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LatestUpdate"
  /discover/top-authors:
    get:
      operationId: getTopAuthors
      tags: [discover]
      summary: Get top authors by followers
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top authors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        displayName:
          type: string
        role:
          type: string
          enum: [author, reader]
          default: reader
        bio:
          type: string
      required:
        - username
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        role:
          type: string
          enum: [author, reader, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - role
        - createdAt
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        role:
          type: string
        seriesCount:
          type: integer
        followersCount:
          type: integer
        followingCount:
          type: integer
        totalViews:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - role
        - seriesCount
        - followersCount
        - followingCount
        - totalViews
        - createdAt
...
    Series:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        coverImage:
          type: string
          nullable: true
        bannerImage:
          type: string
          nullable: true
        type:
          type: string
          enum: [manga, webtoon, comic]
        status:
          type: string
          enum: [ongoing, completed, hiatus]
        genres:
          type: array
          items:
            type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: string
          nullable: true
        chapterCount:
          type: integer
        viewCount:
          type: integer
        likeCount:
          type: integer
        favoriteCount:
          type: integer
        mature:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - type
        - status
        - authorId
        - authorName
        - chapterCount
        - viewCount
        - likeCount
        - createdAt
        - updatedAt
...
    Chapter:
      type: object
      properties:
        id:
          type: integer
        seriesId:
          type: integer
        number:
          type: number
        title:
          type: string
        pageCount:
          type: integer
        viewCount:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - seriesId
        - number
        - title
        - pageCount
        - viewCount
        - likeCount
        - commentCount
        - createdAt
    ChapterDetail:
      type: object
      properties:
        id:
          type: integer
        seriesId:
          type: integer
        seriesTitle:
          type: string
        number:
          type: number
        title:
          type: string
        pages:
          type: array
          items:
            $ref: "#/components/schemas/Page"
        viewCount:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        previousChapterId:
          type: integer
          nullable: true
        nextChapterId:
          type: integer
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - seriesId
        - seriesTitle
        - number
        - title
        - pages
        - viewCount
        - likeCount
        - commentCount
        - createdAt
    CreateChapterBody:
      type: object
      properties:
        number:
          type: number
        title:
          type: string
        published:
          type: boolean
          default: true
      required:
        - number
        - title
    UpdateChapterBody:
      type: object
      properties:
        title:
          type: string
        published:
          type: boolean
    Page:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        pageNumber:
          type: integer
        imageUrl:
          type: string
        width:
          type: integer
        height:
          type: integer
      required:
        - id
        - chapterId
        - pageNumber
        - imageUrl
    AddPagesBody:
      type: object
      properties:
        pages:
          type: array
          items:
            type: object
            properties:
              pageNumber:
                type: integer
              imageUrl:
                type: string
              width:
                type: integer
              height:
                type: integer
            required:
              - pageNumber
              - imageUrl
      required:
        - pages
...
    LatestUpdate:
      type: object
      properties:
        seriesId:
          type: integer
        seriesTitle:
          type: string
        seriesCover:
          type: string
          nullable: true
        seriesType:
          type: string
        chapterId:
          type: integer
        chapterNumber:
          type: number
        chapterTitle:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
      required:
        - seriesId
        - seriesTitle
        - seriesType
        - chapterId
        - chapterNumber
        - chapterTitle
        - authorName
        - publishedAt
...