# lib/api-zod/src/generated/types/createChapterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MangaGramm API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateChapterBody {
  number: number;
  title: string;
  published?: boolean;
}